feat(rust): split xx key agreement into multiple modules and structs
