+++ implementations/rust/ockam/ockam_key_exchange_xx/CHANGELOG.md
@@ -1,0 +1,19 @@
+# Changelog
+
+All notable changes to this crate will be documented in this file.
+
+The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## v0.2.0 - 2021-02-16
+### Added
+- Symmetric and Asymmetric SoftwareVault implementations.
+
+### Modified
+- Updated dependencies.
+- Fixes to error propagation and entity construction.
+
+## v0.1.0 - 2021-02-10
+
+Initial release.
+
